requirements.txt
# Core Web Framework
fastapi
uvicorn[standard]
python-multipart
orjson  # 2-3x faster JSON encoding/decoding vs stdlib json
pydantic-settings  # Pydantic v2 settings management

# YOLO and Triton
ultralytics
tritonclient[all]

# Model Export Dependencies
onnx>=1.12.0,<=1.19.1
onnxslim>=0.1.71
onnxsim>=0.4.33
onnxscript
onnxruntime-gpu
onnx-graphsurgeon  # Required for End2End TensorRT NMS graph manipulation
tensorrt-cu12==10.13.3.9  # MUST match Triton 25.10 TensorRT version (10.13.3.9) for engine compatibility

# NVIDIA DALI (GPU-accelerated preprocessing for Track D)
# Requires NVIDIA PyPI index - see Dockerfile for installation
nvidia-dali-cuda120

# PyTorch (for Track A baseline)
torch
torchvision

# Image Processing
opencv-python
# Note: Pillow is installed as dependency of ultralytics (no need to specify)

# Scientific Computing
numpy
scipy

# HTTP Client
requests
aiohttp
httpx

# Benchmarking and Visualization
matplotlib
seaborn
pandas
psutil
tqdm
tabulate

# Configuration
pyyaml

# Track E: OpenSearch for Visual Search
opensearch-py>=2.3.0  # Async OpenSearch client with k-NN support
transformers>=4.30.0  # CLIP tokenizer for text-to-image search

# Track E: MobileCLIP dependencies
timm  # pytorch-image-models - required for MobileCLIP architecture
huggingface_hub  # For downloading MobileCLIP checkpoints
ftfy  # Text cleaning for CLIP tokenizer (required by open_clip)
